+++ src/arrOfObjs.js
@@ -1,0 +1,32 @@
+const arrOfObjs = [
+  {
+    name: 'Name',
+    score: '100',
+  },
+  {
+    name: 'Name',
+    score: '20',
+  },
+  {
+    name: 'Name',
+    score: '50',
+  },
+  {
+    name: 'Name',
+    score: '78',
+  },
+  {
+    name: 'Name',
+    score: '125',
+  },
+  {
+    name: 'Name',
+    score: '77',
+  },
+  {
+    name: 'Name',
+    score: '42',
+  },
+];
+
+export default arrOfObjs;
